requirements.txt
discord
google
mysql-connector-python